fix(material/snack-bar): snack-bar action button does not follow material guidelines
Fixes a bug in the Angular Material snack-bar, where accent color is used for action
button color in light themes and grey is used in dark themes. This does not follow the
material design guidelines which recommend a tone of primary color is used in both light
and dark themes, and that the color should contrast with the background.

Fixes #13905